refactored manage component into sitewide component
