+++ src/app/app.app-routing.module.ts
@@ -5,7 +5,6 @@ import { AboutComponent }       from './about/about.component';
 import { ConnectComponent }     from './connect/connect.component';
 import { R2d2Component }     from './r2d2/r2d2.component';
 import { NotFoundComponent }    from './notfound/notfound.component';
-import { AuthenticationComponent }    from './authentication/authentication.component';
 
 const routes: Routes = [
   { path: '', component: HomeComponent },
@@ -13,7 +12,6 @@ const routes: Routes = [
   { path: 'connect', component: ConnectComponent },
   { path: 'r2d2', component: R2d2Component },
   { path: '404', component: NotFoundComponent },
-  { path: 'auth', component: AuthenticationComponent },
   { path: '**', redirectTo: '/404' }
 ];
