+++ src/app/blog/blog.module.ts
@@ -3,7 +3,6 @@ import { CommonModule } from '@angular/common';
 import { BlogComponent } from './blog.component';
 import { BlogRoutingModule } from './blog.routing.module';
 import { ListComponent } from './list/list.component';
-import { ManageComponent } from './manage/manage.component';
 import { AuthGuard } from '../services/auth/auth-guard';
 
 @NgModule({
@@ -13,8 +12,7 @@ import { AuthGuard } from '../services/auth/auth-guard';
   ],
   declarations: [
     BlogComponent,
-    ListComponent,
-    ManageComponent
+    ListComponent
   ],
   providers: [
     AuthGuard
