+++ src/app/blog/blog.routing.module.ts
@@ -3,8 +3,6 @@ import { RouterModule, Routes } from '@angular/router';
 
 import { BlogComponent } from './blog.component';
 import { ListComponent } from './list/list.component';
-import { ManageComponent } from './manage/manage.component';
-import { AuthGuard } from '../services/auth/auth-guard';
 
 const blogRoutes: Routes = [
   {
@@ -14,13 +12,6 @@ const blogRoutes: Routes = [
       {
         path: '',
         component: ListComponent
-      },
-      {
-        path: 'manage',
-        component: ManageComponent,
-        canActivate: [
-          AuthGuard
-        ]
       }
     ]
   },
